requirements.txt
selenium
pytest
allure-pytest
faker
webdriver-manager
